UI: split into screens, views and components
