+++ app/src/androidTest/java/io/github/stscoundrel/polylinguist/ui/CurrentStatisticsViewModelTest.kt
@@ -5,6 +5,7 @@ import io.github.stscoundrel.polylinguist.domain.usecase.GetCurrentStatisticsUse
 import io.github.stscoundrel.polylinguist.domain.usecase.GetLatestStatisticsUseCase
 import io.github.stscoundrel.polylinguist.testdata.InMemoryStatisticsRepository
 import io.github.stscoundrel.polylinguist.testdata.StatisticFactory
+import io.github.stscoundrel.polylinguist.ui.viewmodels.CurrentStatisticsViewModel
 import junit.framework.TestCase.assertEquals
 import junit.framework.TestCase.assertFalse
 import junit.framework.TestCase.assertTrue
